Re-organize using containers.
