+++ src/app/containers/login/actions/spec.js
