+++ src/app/containers/login/reducers/spec.js
