+++ src/app/epics.js
@@ -1,5 +1,6 @@
 import {combineEpics} from 'redux-observable';
 
-import {loginEpics} from './components/login/actions';
+import {loginEpics} from './containers/login/actions';
+import {recognitionsEpics} from './containers/header/components/recognitions/epics';
 
-export default combineEpics(loginEpics);
+export default combineEpics(loginEpics, recognitionsEpics);
